Edge cases for Liq Price
Also noting problem in rPNL due to non-integration of clearing with rPNL, so the reduce margin isn't returning the margin, the avg. price ends up null when the new position is opened, leading to Infinity liq price for one side. Avg. price considered for rPNL is the trade price not the position's earlier Avg. entry price.