uPNL included in Liq calc
Good for not sliding the position closer when you update. Now any loss from an existing position (but not gain which is already included in the available, is that right??) is going to factor in which should make the price stay still/accurate when adding to a loser for instance.