Issue is
3 rpnl for liq. counterparty total to an excess.

Liq. address debits deleverage, then the orders.

Liq. counterparty debits orders, then there's a reset of the liq address of the order balances... and then for the counterparty, so it's just the half off no liq.

Then there's a bad rounding of this errant adjusted value and we end up with 11.

3625912 is the block. It's a partial liq that must become a full liq, so this edge case needs to be sanded down.